Load catalogs using the importing ontology's URI scheme
This changes OmlRead.findCatalogs() to load discovered catalog files using the
importing ontology's URI scheme rather than always as a file:// URI. The search
process is otherwise unchanged and requires that the importing ontology URI be
convertible to a file:// URI.

The previous behavior was to convert the importing ontology's URI to a file://
URI, and then look for and load the catalog as a file:// URI, which would result
in rewriteURIs with relative rewritePrefixes also resolving to file:// URIs.
This can cause problems in Eclipse as files opened for editing are loaded with
platform:// URIs; if imported ontologies are loaded with file:// URIs this
inconsistency can result in changes made to imported ontologies not being
reflected in editors for the importing ontology. With this change, the catalog
would be loaded with and resolve relative rewritePrefixes to platform:// URIs.