Fix OmlRead.findCatalogs() in Eclipse on Windows (#58)
To convert Eclipse platform URLs to File objects, we were passing the converted file URLs from
FileLocator.toFileURL() to Paths.get() and then converting the Path to a File, but on Windows
the file URLs FileLocator.toFileURL() returns are not accepted by Paths.get(). Instead, we now
use the file URL path string to construct the File.